pyproject.toml
[project]
name = "pdf-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "aioredis>=2.0.1",
    "asyncpg>=0.31.0",
    "boto3>=1.42.49",
    "celery>=5.6.2",
    "docling>=2.73.1",
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.129.0",
    "greenlet>=3.3.1",
    "msgspec>=0.20.0",
    "omegaconf>=2.3.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "passlib>=1.7.4",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13.0",
    "python-jose>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-magic>=0.4.27",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "pyrefly>=0.52.0",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"


# Pyrefly header
[tool.pyrefly]

#### configuring what to type check and where to import from
project-includes = ["src", "."]
project-excludes = ["**/.[!/.]*"]
search-path = ["."]
